Exmaple Website Style/requirements.txt
Django==4.2.13
Pillow>=10.0.0
python-decouple==3.8
requests==2.31.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-widget-tweaks==1.4.12